Resynced with GNU Classpath: getLayer fixes in Swing
2004-12-14  Dalibor Topic  <robilad@kaffe.org>

        * libraries/javalib/javax/swing/JLayeredPane.java,
        libraries/javalib/javax/swing/JInternalFrame.java:
        Resynced with GNU Classpath.

        2004-12-13  Michael Koch  <konqueror@gmx.de>

        * javax/swing/JLayeredPane.java
        (getLayer): Fixed return type, made public.
        (getPosition): Use new getLayer().
        (SetPosition): Likewise.
        (getIndexOf): Likewise.
        (remove): Likewise.
        * javax/swing/JInternalFrame.java
        (getLayer): Use new JLayeredPane.getLayer().

+++ libraries/javalib/javax/swing/JInternalFrame.java
@@ -860,7 +860,7 @@ public class JInternalFrame extends JComponent implements Accessible,
   {
     JDesktopPane pane = getDesktopPane();
     if (pane != null)
-      return pane.getLayer(this).intValue();
+      return pane.getLayer(this);
     return -1;
   }
